refactor: ResultView 제거 후 OutputView에서 구현
